docker-consumer/requirements.txt
pika==1.3.2
pymongo==4.6.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pydantic==2.10.13
pydantic-settings==2.3.4